Subscription features implemented. Locked featrues, free features

+++ src/app/modules/barberSchedule/barberSchedule.controller.ts
@@ -2,16 +2,35 @@ import httpStatus from 'http-status';
 import sendResponse from '../../utils/sendResponse';
 import catchAsync from '../../utils/catchAsync';
 import { barberScheduleService } from './barberSchedule.service';
+import { SubscriptionPlanStatus } from '@prisma/client';
 
 const createBarberSchedule = catchAsync(async (req, res) => {
   const user = req.user as any;
-  const result = await barberScheduleService.createBarberScheduleIntoDb(user.id, req.body);
-  sendResponse(res, {
-    statusCode: httpStatus.CREATED,
-    success: true,
-    message: 'BarberSchedule created successfully',
-    data: result,
-  });
+  const subscriptionPlanName = user.subscriptionPlan;
+  if (
+    subscriptionPlanName === SubscriptionPlanStatus.FREE ||
+    subscriptionPlanName === SubscriptionPlanStatus.ADVANCED_PREMIUM
+  ) {
+    return sendResponse(res, {
+      statusCode: httpStatus.FORBIDDEN,
+      success: false,
+      message:
+        'Access denied. Upgrade your subscription to access to manage barber schedules.',
+      data: null,
+    });
+  }
+  if (
+    subscriptionPlanName === SubscriptionPlanStatus.BASIC_PREMIUM ||
+    SubscriptionPlanStatus.PRO_PREMIUM
+  ) {
+    const result = await barberScheduleService.createBarberScheduleIntoDb(user.id, req.body);
+    sendResponse(res, {
+      statusCode: httpStatus.CREATED,
+      success: true,
+      message: 'BarberSchedule created successfully',
+      data: result,
+    });
+  }
 });
 
 const getBarberScheduleList = catchAsync(async (req, res) => {
@@ -38,24 +57,60 @@ const getBarberScheduleById = catchAsync(async (req, res) => {
 
 const updateBarberSchedule = catchAsync(async (req, res) => {
   const user = req.user as any;
-  const result = await barberScheduleService.updateBarberScheduleIntoDb(user.id, req.params.id, req.body);
-  sendResponse(res, {
-    statusCode: httpStatus.OK,
-    success: true,
-    message: 'BarberSchedule updated successfully',
-    data: result,
-  });
+  const subscriptionPlanName = user.subscriptionPlan;
+  if (
+    subscriptionPlanName === SubscriptionPlanStatus.FREE ||
+    subscriptionPlanName === SubscriptionPlanStatus.ADVANCED_PREMIUM
+  ) {
+    return sendResponse(res, {
+      statusCode: httpStatus.FORBIDDEN,
+      success: false,
+      message:
+        'Access denied. Upgrade your subscription to access to manage barber schedules.',
+      data: null,
+    });
+  }
+  if (
+    subscriptionPlanName === SubscriptionPlanStatus.BASIC_PREMIUM ||
+    SubscriptionPlanStatus.PRO_PREMIUM
+  ) {
+    const result = await barberScheduleService.updateBarberScheduleIntoDb(user.id, req.params.id, req.body);
+    sendResponse(res, {
+      statusCode: httpStatus.OK,
+      success: true,
+      message: 'BarberSchedule updated successfully',
+      data: result,
+    });
+  }
 });
 
 const deleteBarberSchedule = catchAsync(async (req, res) => {
   const user = req.user as any;
-  const result = await barberScheduleService.deleteBarberScheduleItemFromDb(user.id, req.params.id);
-  sendResponse(res, {
-    statusCode: httpStatus.OK,
-    success: true,
-    message: 'BarberSchedule deleted successfully',
-    data: result,
-  });
+  const subscriptionPlanName = user.subscriptionPlan;
+  if (
+    subscriptionPlanName === SubscriptionPlanStatus.FREE ||
+    subscriptionPlanName === SubscriptionPlanStatus.ADVANCED_PREMIUM
+  ) {
+    return sendResponse(res, {
+      statusCode: httpStatus.FORBIDDEN,
+      success: false,
+      message:
+        'Access denied. Upgrade your subscription to access to manage barber schedules.',
+      data: null,
+    });
+  }
+  if (
+    subscriptionPlanName === SubscriptionPlanStatus.BASIC_PREMIUM ||
+    SubscriptionPlanStatus.PRO_PREMIUM
+  ) {
+    const result = await barberScheduleService.deleteBarberScheduleItemFromDb(user.id, req.params.id);
+    sendResponse(res, {
+      statusCode: httpStatus.OK,
+      success: true,
+      message: 'BarberSchedule deleted successfully',
+      data: result,
+    });
+  }
 });
 
 export const barberScheduleController = {
